[Bronze II] Title: 숫자, Time: 16 ms, Memory: 3684 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -5,15 +5,15 @@ using namespace std;
 int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(0); cout.tie(0);
-    int a, b, cnt = 0;
+    long long a, b, cnt = 0;
     cin >> a >> b;
     if(a>b){
-        int tmp = b;
+        long long tmp = b;
         b = a;
         a = tmp;
     }
-    vector<int>v;
-    for(int i=a+1;i<b;i++){
+    vector<long long>v;
+    for(long long i=a+1;i<b;i++){
         v.push_back(i);
         cnt++;
     }
